Fix the buffer names for SQLOut and SQLCatalog.
bufnr() uses a case-insensitive regex search of buffers, which would
find files like sqlcatalog.vim instead of SQLCatalog. Using a string
like '^SQLCatalog$' fixes the search, but also creates a buffer with the
caret and dollar sign in its name. I decided to fix that by putting
uncommon characters in the buffer name.