Improve the Action queries' syntax and formatting.
Use QUOTENAME and QUOTE_IDENT instead of concatenating the brackets or
quotation marks. Also, don't do this more than necessary. Note: this
doesn't catch all cases; SQLServer keywords won't be quoted by this
logic, but that shouldn't happen often.

Used sql-formatter to tidy up the appearance of the query. I'm not 100%
satisfied with the results, but it does look better.