Implements #9. From the Actions pop-up menu, press `L`.
This will create a new SQL buffer in the same window as the original SQL
buffer. Its name is a concatenation of the action taken (Script, Select,
Insert, etc.), the database name, the schema and object name, and a
`sql` extension.

If the file was already generated and modified, repeating this process
will refresh the buffer, wiping away any edits.